+++ requirements/common.txt
@@ -8,7 +8,7 @@
 # with "TypeError: unsupported operand type(s) for |: 'list' and 'set'".
 # That breakage is now handled in code instead of by pinning:
 # MossTTSNanoForGeneration wraps weight loading in
-# _transformers_keys_to_ignore_compat (modeling_moss_tts_nano.py).
+# transformers_keys_to_ignore_compat (vllm_omni/model_executor/models/utils.py).
 transformers >= 4.56.0, != 5.0.*, != 5.1.*, != 5.2.*, != 5.3.*, != 5.4.*, != 5.5.0
 av>=14.0.0
 omegaconf>=2.3.0
